requirements.txt
python-telegram-bot==20.8
keepa==1.3.0
SQLAlchemy==2.0.0
requests==2.31.0
pillow==10.2.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3